Ignore Pattern (#17)
This adds an ignore pattern to skip some executable files from the init and cleanup directories, if necessary. The default is to ignore `*.md`, in order to support the special case of [WSL](https://docs.microsoft.com/en-us/windows/wsl/install) as all files from Windows partitions are mounted with the `x` flag by default.