compose shim (#23)
This implements an (almost) drop-in replacement for `docker-compose build`. The shim, called `compose.sh` can be installed at the OS level as a `docker-compose` and will permit building of images from external tools, as it supports a (large) subset of the CLI options and environment variables of `docker-compose` and its `build` command. The aim of this shim is to facilitate building Docker images without the `docker` client, as long as the client provided supports the same set of options, e.g. `img` or `nerdctl`.